Utilising a controller for strava related requests
